+++ src/main/java/com/example/gistcompetitioncnserver/comment/application/PostDeleteEventListenerOfComment.java
@@ -1,7 +1,7 @@
 package com.example.gistcompetitioncnserver.comment.application;
 
 import com.example.gistcompetitioncnserver.comment.domain.CommentRepository;
-import com.example.gistcompetitioncnserver.post.PostDeleteEvent;
+import com.example.gistcompetitioncnserver.post.application.PostDeleteEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
