+++ src/test/java/com/example/gistcompetitioncnserver/answer/AnswerServiceTest.java
@@ -8,8 +8,8 @@ import com.example.gistcompetitioncnserver.answer.dto.AnswerRequest;
 import com.example.gistcompetitioncnserver.exception.WrappedException;
 import com.example.gistcompetitioncnserver.exception.post.NoSuchPostException;
 import com.example.gistcompetitioncnserver.exception.post.UnAnsweredPostException;
-import com.example.gistcompetitioncnserver.post.Post;
-import com.example.gistcompetitioncnserver.post.PostRepository;
+import com.example.gistcompetitioncnserver.post.domain.Post;
+import com.example.gistcompetitioncnserver.post.domain.PostRepository;
 import com.example.gistcompetitioncnserver.user.User;
 import com.example.gistcompetitioncnserver.user.UserRepository;
 import com.example.gistcompetitioncnserver.user.UserRole;
