+++ src/test/java/com/example/gistcompetitioncnserver/post/PostTest.java
@@ -1,7 +1,7 @@
 package com.example.gistcompetitioncnserver.post;
 
 import com.example.gistcompetitioncnserver.exception.post.DuplicatedAgreementException;
-import com.example.gistcompetitioncnserver.exception.post.PostException;
+import com.example.gistcompetitioncnserver.post.domain.Post;
 import com.example.gistcompetitioncnserver.user.User;
 import com.example.gistcompetitioncnserver.user.UserRole;
 import org.junit.jupiter.api.BeforeEach;
